# dist/openapi-complete.yaml
openapi: 3.0.0
info:
  title: המתכונים של סבתא ואחרים - Grandma's Recipes and Others API
  description: Complete API for recipes project including authentication, search, personal area, and bonus features
  version: 2.0.0
servers:
  - url: "http://127.0.0.1:80"
    description: Local development server
...
paths:
...
  /register:
    post:
      tags:
        - Auth
      summary: רישום משתמש חדש
      description: A JSON object contains all the details about the new user.
      requestBody:
        description: User registration data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "201":
          description: A new user has been added
        "400":
          description: wrong input parameters
  /login:
    post:
      tags:
        - Auth
      summary: התחברות משתמש קיים
      description: An exist user login to the system
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginScheme"
        required: true
      responses:
        "200":
          description: User authenticated successfully
          headers:
            Set-Cookie:
              style: simple
              explode: false
              schema:
                type: string
                example: session=87s97sdhgf279293f8w0e9
        "401":
          description: Authentication failed
...
  /recipes/search:
    get:
      tags:
        - Recipes
      summary: חיפוש מתכונים
      parameters:
        - name: recipeName
          in: query
          schema:
            type: string
        - name: cuisine
          in: query
          schema:
            type: string
        - name: diet
          in: query
          schema:
            type: string
        - name: intolerance
          in: query
          schema:
            type: string
        - name: number
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecipePreview"
  /recipes/random:
    get:
      tags:
        - Recipes
      summary: קבלת מתכונים אקראיים
      parameters:
        - name: number
          in: query
          schema:
            type: integer
            default: 3
      responses:
        "200":
          description: Random recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecipePreview"
  /recipes/{recipeId}:
    get:
      tags:
        - Recipes
      summary: קבלת פרטי מתכון
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recipe details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeDetails"
        "404":
          description: Recipe not found
  /recipes/{recipeId}/instructions:
    get:
      tags:
        - Recipes
      summary: קבלת הוראות הכנה
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recipe instructions

  # Filter Endpoints
...
  /users/{userId}/recipes/{recipeId}:
    delete:
      tags:
        - Personal
      summary: מחיקת מתכון אישי
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: recipeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recipe deleted successfully
      security:
        - cookieAuth: []
...
components:
  schemas:
    User:
      type: object
      required:
        - username
        - firstname
        - lastname
        - country
        - password
        - email
      properties:
        username:
          maxLength: 8
          minLength: 3
          type: string
          description: username contains between 3 to 8 letters only
          example: my_name
        firstname:
          type: string
          description: user's first name
          example: my_first_name
        lastname:
          type: string
          description: user's last name
          example: my_last_name
        country:
          type: string
          description: user's country
          example: Israel
        password:
          maxLength: 10
          minLength: 5
          type: string
          description: password contains between 5 to 10 chars, at least one number and special char
          format: password
          example: my_name@12
        email:
          type: string
          format: email
          description: user's email
          example: my_name12@gmail.com
    LoginScheme:
      type: object
      required:
        - username
        - password
      properties:
        username:
          maxLength: 8
          minLength: 3
          type: string
          description: username contains letters only
        password:
          maxLength: 10
          minLength: 5
          description: password contains at least one number and special char
          format: password
      example:
        username: my_name
        password: my_name@12
    RecipePreview:
      type: object
      properties:
        id:
          type: integer
          description: recipe identifier
          example: 102
        title:
          type: string
          description: recipe name
          example: Simple Macaroni and Cheese
        image:
          type: string
          description: recipe image
          format: URL
          example: https://spoonacular.com/recipeimages/655455-566x370.jpg
        readyInMinutes:
          type: integer
          description: time to make this recipe
          example: 45
        aggregateLikes:
          type: integer
          description: how many likes this recipe got
          example: 5987
        vegetarian:
          type: boolean
          description: indication if the recipe is vegetarian
          example: false
        vegan:
          type: boolean
          description: indication if the recipe is vegan
          example: false
        glutenFree:
          type: boolean
          description: indication if the recipe is gluten free
          example: false
    RecipeDetails:
      allOf:
        - $ref: "#/components/schemas/RecipePreview"
        - type: object
          properties:
            summary:
              type: string
              description: recipe summary
            servings:
              type: integer
              description: number of servings
            extendedIngredients:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                  original:
                    type: string
                  amount:
                    type: number
                  unit:
                    type: string
            analyzedInstructions:
              type: array
              items:
                type: object
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session